支持发送上传图片且默认原图

+++ Ore-Base/src/main/kotlin/moe/ore/helper/NumberExt.kt
@@ -13,4 +13,4 @@ fun Int.toHexString(): String = Integer.toHexString(this)
 
 fun Int.intToIp(reverse: Boolean = false): String = if(reverse) IpUtil.int_to_ip(this.toLong()) else IpUtil.getNumConvertIp(this.toLong())
 
-fun Long.longToIp(reverse: Boolean = false): String = if(reverse) IpUtil.int_to_ip(this.toLong()) else IpUtil.getNumConvertIp(this.toLong())
+fun Long.longToIp(reverse: Boolean = false): String = if(reverse) IpUtil.int_to_ip(this) else IpUtil.getNumConvertIp(this)
